Don't duplicate a link path.
Somehow with `path '/holdings/:holding_id'`, `link` is malformed to be `localhost…/holdings/1234/holdings/1234`.

This is because inside `lib/flappi/response_builder.rb#controller_base_url`, `source_definition.endpoint_info[:path]` is mutated to be `/holdings/1234`, not `/holdings/:holding_id`.

I couldnt' find the source of the mutation, so I decided to give a very rough example of it wrongly "fixed" in this PR.